+++ CHANGES.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ## Fixed
 - Instagram Crawler
+- Telegram Channel username/title
 
 [1.0.0] - 2021-07-13
 ### Added